[BUG] [GCC] warning: comparison of integer expressions 5

+++ src/common/scripting/vm/vm.h
@@ -738,7 +738,7 @@ typename VMReturnTypeTrait<RetVal>::type VMCallSingle(VMFunction* func, Args... 
 
 		VMValue params[argCount];
 
-		for(int i = 0, j = 0; i < sizeof...(Args); i++)
+		for(size_t i = 0, j = 0; i < sizeof...(Args); i++)
 		{
 			for(int k = 0; k < arglist[i].count; k++, j++)
 			{
@@ -774,7 +774,7 @@ std::tuple<typename VMReturnTypeTrait<Rets>::type...> VMCallMultiImpl(VMFunction
 
 		VMValue params[argCount];
 
-		for(int i = 0, j = 0; i < sizeof...(Args); i++)
+		for(size_t i = 0, j = 0; i < sizeof...(Args); i++)
 		{
 			for(int k = 0; k < arglist[i].count; k++, j++)
 			{
