+++ src/playsim/p_mobj.cpp
@@ -1374,7 +1374,7 @@ DEFINE_ACTION_FUNCTION(AActor, CheckLocalView)
 
 void AActor::DisableLocalRendering(const unsigned int pNum, const bool disable)
 {
-	if (pNum == consoleplayer)
+	if ((int)pNum == consoleplayer)
 		NoLocalRender = disable;
 }
